alpha
Auto synchronizing

+++ webpack.config.js
@@ -22,7 +22,8 @@ let entry = glob
     return entries;
   }, {});
 
-const isDev = true; //env.NODE_ENV === 'development';
+const isDev = false; //env.NODE_ENV === 'development';
+const isDevServer = !!process.env.WEBPACK_DEV_SERVER;
 module.exports = (env, argv) => {
   return {
     mode: isDev ? 'development' : 'production',
@@ -166,6 +167,8 @@ module.exports = (env, argv) => {
     plugins: [
       new WebpackUserscript({
         headers: function (data) {
+          
+          console.log(isDevServer)
           let origionpath = entry[data.chunkName];
 
           if (!fs.existsSync(origionpath)) {
@@ -183,7 +186,7 @@ module.exports = (env, argv) => {
               [data.chunkHash]: vstring
             };
 
-            if (!fs.existsSync(versionpath)) {
+            if (!isDevServer && !fs.existsSync(versionpath)) {
               fs.writeFileSync(versionpath, curVersionJson);
             }
 
@@ -201,7 +204,8 @@ module.exports = (env, argv) => {
               });
             } else {
               //hash不同
-              fs.writeFileSync(versionpath, JSON.stringify(curVersionJson), 'utf8');
+              if (!isDevServer)
+                fs.writeFileSync(versionpath, JSON.stringify(curVersionJson), 'utf8');
               return extend(true, {}, header, {
                 version: curVersionJson[data.chunkHash]
               });
@@ -222,7 +226,7 @@ module.exports = (env, argv) => {
       })
     ],
     devServer: {
-      writeToDisk: true,
+
       publicPath: '/',
       contentBase: path.join(__dirname, 'dist')
     },
